python-base: also fix the right in the code dir
