+++ python_base/Dockerfile
@@ -88,8 +88,8 @@ RUN useradd test
 RUN chown -R test:test /code /deps /src-cache /pip-cache
 # This allows us to change the venv rights after it has been created
 ADD ./python_base/docker_entrypoint.sh /docker_entrypoint.sh
-ADD ./python_base/fix_venv_rights /fix_venv_rights
-RUN chmod 4755 /fix_venv_rights
+ADD ./python_base/fix_rights /fix_rights
+RUN chmod 4755 /fix_rights
 
 USER test
 WORKDIR /code
